+++ backend/apps/iamstudent/filters.py
@@ -3,7 +3,8 @@ from django.utils.translation import gettext_lazy as _
 import django_filters as filters
 
 from .forms import get_form_helper_filter
-from .models import AUSBILDUNGS_TYPEN_COLUMNS, BEZAHLUNG_CHOICES_Filter, Student
+from .models import Student
+from .models.student import AUSBILDUNGS_TYPEN_COLUMNS, BEZAHLUNG_CHOICES_Filter
 
 
 class StudentJobRequirementsFilter(filters.FilterSet):
