+++ backend/apps/iamstudent/forms_filter.py
@@ -2,7 +2,7 @@ from django import forms
 from django.utils.translation import gettext_lazy as _
 
 from .forms import form_labels
-from .models_persistent_filter import StudentListFilterModel
+from .models import StudentListFilterModel
 
 
 class StudentListFilterModelForm(forms.ModelForm):
