+++ backend/apps/iamstudent/models/persistent_filter.py
@@ -5,11 +5,10 @@ from django.db import models
 import django.forms as forms
 import django_filters.fields as filter_fields
 
+from apps.iamstudent.filters import StudentJobRequirementsFilter
 from apps.ineedstudent.models import Hospital
 
-from .filters import StudentJobRequirementsFilter
-from .models import *  # noqa: F401, F403
-from .models import COUNTRY_CODE_CHOICES
+from .student import COUNTRY_CODE_CHOICES
 
 
 class LocationFilterModel(models.Model):
